+++ lib/wits.rb
@@ -2,6 +2,7 @@ require 'faraday'
 
 require 'wits/version'
 require 'wits/nodes'
+require 'wits/price_codes'
 require 'wits/final_interim_prices'
 require 'wits/five_minute_prices'