+++ lib/wits/final_interim_prices/convenience_methods.rb
@@ -24,6 +24,16 @@ module Wits
 
           # def halfway_bush_prices
           create_prices_method("#{name.downcase.tr(' ', '_')}_prices", node)
+
+
+          # def ben2201_interim_prices
+          create_interim_prices_method(node.downcase, node)
+
+          # def hay_interim_prices
+          create_interim_prices_method(node.downcase[0..2], node)
+
+          # def halfway_bush_interim_prices
+          create_interim_prices_method(name.downcase.tr(' ', '_'), node)
         end
       end
 
@@ -32,6 +42,12 @@ module Wits
       def self.create_prices_method(method_name, node)
         define_method method_name do |*args|
           prices(node, *args)
+        end
+      end
+
+      def self.create_interim_prices_method(method_prefix, node)
+        define_method "#{method_prefix}_interim_prices" do |*args|
+          interim_prices(node, *args)
         end
       end
     end