+++ lib/wits/five_minute_prices.rb
@@ -10,82 +10,20 @@ module Wits
     extend self
     extend Helpers
     extend ConvenienceMethods
+    include PriceCodes
 
     def self.extended(base)
       base.extend Helpers
       base.extend ConvenienceMethods
     end
 
+    def five_minute_prices(node, date = nz_current_date, type = FIVE_MINUTE)
+      csv = request_prices(node, date, type)
+      parse_csv(csv, type)
+    end
+
     def average_five_minute_prices(node, date = nz_current_date)
-      five_minute_prices(node, date, :average)
-    end
-
-    def five_minute_prices(node, date = nz_current_date, type = :price)
-      node = format_node(node)
-      date = format_date(date)
-      type = format_price_type(type)
-
-      request_five_min_prices(node, date, type)
-    end
-
-    private
-
-    def format_price_type(type)
-      if type == :price
-        'Price'
-      else
-        'Average'
-      end
-    end
-
-    def request_five_min_prices(node, date, type)
-      response = Wits::Client.get_csv do |request|
-        request.url(
-          'comitFta/five_min_prices.download',
-          INchoice:    'SEL', # Mandatory
-          INdate:       date,
-          INgip:        node,
-          INperiodfrom: 1,
-          INperiodto:   50,
-          INtype:       type
-        )
-      end
-
-      parse_five_min_csv(response.body, type)
-    end
-
-    def parse_five_min_csv(csv, type)
-      csv = CSV.parse(csv.sub("\r\n", "\n"))
-
-      node, date = process_five_min_metadata(csv)
-      prices     = process_five_min_prices(csv)
-      price_type = 'Five Minute'
-
-      price_type = "Average #{price_type}" if type == 'Average'
-
-      format_prices(node, date, prices, price_type)
-    rescue StandardError => error
-      raise Wits::Error::ParsingError.new(error)
-    end
-
-    def process_five_min_metadata(csv)
-      csv.shift # discard header
-
-      node = csv.first[0]
-      date = Date.parse csv.first[1]
-
-      [node, date]
-    end
-
-    def process_five_min_prices(csv)
-      times = []
-
-      csv.map do |_node, date, trading_period, time, price, *_|
-        repeated_time = times.include?(time)
-        times << time
-
-        format_price(date, time, trading_period, price, repeated_time)
-      end
+      five_minute_prices(node, date, AVERAGE)
     end
   end
 end