+++ spec/spec_helper.rb
@@ -29,7 +29,7 @@ require 'support/request_helpers'
 require 'support/time_parser'
 
 VCR.configure do |config|
-  config.cassette_library_dir = "spec/vcr_cassettes"
+  config.cassette_library_dir = 'spec/vcr_cassettes'
   config.hook_into :webmock
   config.default_cassette_options = { serialize_with: :json }
   config.configure_rspec_metadata!